Use `handle` for handlers with internal state
Because `hmap'` was used in the handlers for `Deferred` and
`Input`, the internal state (i.e., the result of the forced
computation or the remaining input) was reset sometimes when
scoped effects were used. The problem was fixed by using `handle`
instead and test cases were added to ensure that the bug is
fixed.